backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
firebase-admin==6.5.0
pydantic==2.9.2
pydantic-settings==2.6.0

# ML Libraries
scikit-learn==1.5.2
pandas==2.2.3
numpy==2.1.2
scipy==1.14.1

# Database
pymongo==4.10.1
motor==3.6.0

# HTTP Client
httpx==0.27.2
aiohttp==3.10.10

# CORS
python-multipart==0.0.12

# Azure Integration
azure-storage-blob==12.23.1
azure-identity==1.19.0

# OpenAI (for advanced recommendations)
openai==1.54.3

# Caching
redis==5.2.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
